+++ app/serializers/manage/organisation_serializer.rb
@@ -5,7 +5,7 @@ module Manage
     association :booking_categories, blueprint: BookingCategorySerializer
     association :rich_text_templates, blueprint: RichTextTemplateSerializer
     association :homes, blueprint: HomeSerializer
-    association :tenants, blueprint: TenantSerializer
+    # association :tenants, blueprint: TenantSerializer
     association :designated_documents, blueprint: DesignatedDocumentSerializer
     association :tarifs, blueprint: TarifSerializer
     association :booking_categories, blueprint: BookingCategorySerializer