+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Feature: Add soft delete for booking categories
+- Feature: Add custom booking questions
 
 ## Version 23.6.2